+++ packages/react/src/experimental/Forms/Fields/Select/__stories__/index.stories.tsx
@@ -75,7 +75,7 @@ const items = [
   },
   {
     id: "system",
-    name: "System",
+    name: "System with a long label can overflow",
     description: "A theme that adapts to the system's default appearance",
     tag: "Unpaid",
   },