+++ packages/react/src/experimental/OneDataCollection/__stories__/empty-states.stories.tsx
@@ -46,6 +46,10 @@ const BaseExampleComponent = ({
     dataAdapter,
     currentFilters,
     sortings,
+    search: {
+      enabled: true,
+      sync: true,
+    },
   })
 
   const mockVisualizations = getMockVisualizations()