+++ packages/react/src/experimental/OneDataCollection/visualizations/collection/Table/index.tsx
@@ -36,6 +36,7 @@ import { SummariesDefinition, SummaryKey } from "../../../summary"
 import { CollectionProps } from "../../../types"
 import { statusToChecked } from "../utils"
 import { Row } from "./components/Row"
+import { useSticky } from "./useSticky"
 
 export type WithOptionalSorting<
   R extends RecordType,
@@ -268,6 +269,13 @@ export const TableCollection = <
 
   const skeletonColumns =
     columns.length + (source.itemActions ? 1 : 0) + (source.selectable ? 1 : 0)
+
+  const { getStickyPosition, checkColumnWidth } = useSticky(
+    frozenColumnsLeft,
+    columns,
+    !!source.selectable
+  )
+
   /*
    * Initial loading
    */
@@ -286,10 +294,8 @@ export const TableCollection = <
     })
   }
 
-  const checkColumnWidth = source.selectable ? 52 : 0
-
   return (
-    <div className="flex h-full min-h-0 flex-col gap-4">
+    <div className="test flex h-full min-h-0 flex-col gap-4">
       <OneTable loading={isLoading}>
         <TableHeader sticky={true}>
           <TableRow>
@@ -321,18 +327,7 @@ export const TableCollection = <
                 )}
                 width={column.width}
                 align={column.align}
-                sticky={
-                  index < frozenColumnsLeft
-                    ? {
-                        left: columns
-                          .slice(0, Math.max(0, index))
-                          .reduce(
-                            (acc, column) => acc + (column.width ?? 0),
-                            checkColumnWidth
-                          ),
-                      }
-                    : undefined
-                }
+                sticky={getStickyPosition(index)}
                 {...column}
                 onSortClick={
                   sorting
@@ -470,9 +465,9 @@ export const TableCollection = <
             paginationInfo.hasMore && (
               <tr>
                 <td
-                  colSpan={columns.length + (source.selectable ? 1 : 0)}
+                  colSpan={columns.length + (source.selectable ? 1 : 0) + 1}
                   ref={loadingIndicatorRef}
-                  className="h-10 w-full"
+                  className="h-10"
                   aria-hidden="true"
                 ></td>
               </tr>
@@ -502,18 +497,7 @@ export const TableCollection = <
                   key={`summary-${String(column.label)}`}
                   firstCell={cellIndex === 0}
                   width={column.width}
-                  sticky={
-                    cellIndex < frozenColumnsLeft
-                      ? {
-                          left: columns
-                            .slice(0, Math.max(0, cellIndex))
-                            .reduce(
-                              (acc, column) => acc + (column.width ?? 0),
-                              checkColumnWidth
-                            ),
-                        }
-                      : undefined
-                  }
+                  sticky={getStickyPosition(cellIndex)}
                 >
                   {cellIndex === 0 &&
                   !source.selectable &&